DigitalEstate_Agency/my_quart_app/requirements.txt
quart==0.20.0
pymongo==4.5.0    # если используете MongoDB
python-dotenv==1.0.0
requests==2.32.3
motor==3.5.3 
# Далее вы можете добавить библиотеки для векторной БД, LLM и т.д.

